Let the mixer's effect band follow the pointer
It was the one drag area in dew that did not. The height was stored as a number
of knob rows and the drag divided the pointer's travel by size::knobRow and
rounded, so the band had four reachable positions: half a rung of travel changed
nothing and then it jumped a whole one. It was also wrong on its own terms - a
row costs knobRow PLUS space::sm inside a card, so the rounding drag fell six
pixels behind the pointer for every rung it climbed.

The height is a pixel count now and the cards keep their whole rows: the band
follows the pointer, and knobRowsFitting decides how many rows the height it has
landed on can hold. The few pixels between one rung and the next are band ground
rather than a knob standing on part of a row, which is what the old comment was
actually protecting and is still true.

bandHeightForRows becomes public, because whoever owns the height needs both
directions of that formula. Settings moves to "mixerEffectBandHeight" rather
than reinterpreting the old rows key, whose values are one to four and would
open the band at four pixels for anyone upgrading.

And the playlist's own lane drag, which had the mirror-image defect: deltaY /
lanes was int over int, so the fourth lane's edge moved in 4px steps and the
tenth in 10px ones - a resize that got coarser the further down it was grabbed -
while setTrackHeight and RowView::exactHeight both take a double.

+++ src/ui/PlaylistView.cpp
@@ -89,7 +89,11 @@ void PlaylistComponent::dragRowHeightBy (int laneIndex, int deltaY)
     // that follows it, whichever lane it belongs to.
     const auto lanes = juce::jmax (1, laneIndex + 1);
 
-    setTrackHeight (heightAtDragStart + deltaY / lanes);
+    // In DOUBLE, because setTrackHeight and RowView::exactHeight both take one:
+    // as int / int this quantised the drag by the number of lanes above the
+    // grabbed edge, so the fourth lane's edge moved in 4px steps and the tenth
+    // in 10px ones - a resize that got coarser the further down you grabbed it.
+    setTrackHeight (heightAtDragStart + (double) deltaY / lanes);
 }
 
 void PlaylistComponent::endRowHeightDrag()
